# [RBD] Add RBD image performance test for CLAY vs Jerasure EC pools